# server/openapi.yaml
openapi: 3.0.3
info:
  title: Mystická Hvězda API
  version: 1.0.0
  description: |
    REST API for the Mystická Hvězda astrology platform.
    Provides horoscopes, tarot, numerology, oracle readings, and subscription management.

    **Authentication**: JWT token stored in `auth_token` HttpOnly cookie (set on login).
    **CSRF**: All non-GET requests to `/api/*` require an `X-CSRF-Token` header.
    Get a token from `GET /api/csrf-token` (15-minute TTL).
servers:
  - url: /api
    description: API base path (relative)
  - url: https://mystickahvezda.cz/api
    description: Production
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: HttpOnly JWT cookie set on successful login
  schemas:
...
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Invalid email format
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        birth_date:
          type: string
          format: date
        birth_time:
          type: string
          example: "14:30"
        birth_place:
          type: string
        subscription_status:
          type: string
          enum: [free, premium_monthly, exclusive_monthly, vip_majestrat]
        avatar:
          type: string
          nullable: true
...
  parameters:
    XCsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
      description: CSRF token from GET /api/csrf-token (15-minute TTL)
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register new user
      description: |
        Creates a Supabase Auth account. Sends email confirmation.
        User must verify email before logging in.
      parameters:
        - $ref: '#/components/parameters/XCsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, birth_date]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  description: Min 8 chars; must include uppercase, lowercase, digit or special char
                confirm_password:
                  type: string
                first_name:
                  type: string
                birth_date:
                  type: string
                  format: date
                birth_time:
                  type: string
                  example: "14:30"
                birth_place:
                  type: string
      responses:
        "200":
          description: Registration successful — check email for verification link
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  requireEmailVerification:
                    type: boolean
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        "403":
          description: Missing or expired CSRF token
        "429":
          description: Rate limited (10 req/hour per IP)
  /auth/login:
    post:
      tags: [auth]
      summary: Log in
      description: |
        Authenticates via Supabase Auth.
        On success, sets `auth_token` (HttpOnly JWT) and `logged_in` (readable) cookies.
      parameters:
        - $ref: '#/components/parameters/XCsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        "400":
          description: Invalid credentials or unverified email
        "403":
          description: Missing CSRF token
        "429":
          description: Account locked (brute-force protection) or rate limited
  /auth/logout:
    post:
      tags: [auth]
      summary: Log out
      description: Blacklists current JWT and clears auth cookies.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/XCsrfToken'
      responses:
        "200":
          description: Logged out
        "401":
          description: Not authenticated
  /auth/refresh-token:
    post:
      tags: [auth]
      summary: Refresh JWT
      description: Issues a new JWT with up-to-date subscription info.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/XCsrfToken'
      responses:
        "200":
          description: Refreshed token set in cookie
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        "401":
          description: Not authenticated
  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Request password reset email
      description: Always returns success to prevent email enumeration.
      parameters:
        - $ref: '#/components/parameters/XCsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Reset email sent (if account exists)
  /auth/reset-password:
    post:
      tags: [auth]
      summary: Reset password with token
      description: Updates password using Supabase access token from reset email.
      parameters:
        - $ref: '#/components/parameters/XCsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        "200":
          description: Password updated
        "400":
          description: Invalid or expired reset token
  /auth/profile:
    get:
      tags: [auth]
      summary: Get user profile
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Profile with subscription info
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        "401":
          description: Not authenticated
    put:
      tags: [auth]
      summary: Update user profile
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/XCsrfToken'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                birth_date:
                  type: string
                  format: date
                birth_time:
                  type: string
                birth_place:
                  type: string
                avatar:
                  type: string
      responses:
        "200":
          description: Profile updated
        "400":
          description: Validation error
        "401":
          description: Not authenticated

  # ─── Oracle ─────────────────────────────────────────────────────────────────
...